requirements.txt
goopylib~=2.0.0
numpy
pillow
